app/database/base.py
"""Database base configuration."""

import os

from dotenv import load_dotenv
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker

# Load environment variables from .env file
load_dotenv()


def get_database_url() -> str:
    """Get database URL from environment variables."""
    # Check if DATABASE_URL is set (takes precedence)
    database_url = os.getenv("DATABASE_URL")
    if database_url:
        return database_url

    # Otherwise, construct from individual environment variables
    db_host = os.getenv("DB_HOST", "localhost")
    db_port = os.getenv("DB_PORT", "5432")
    db_user = os.getenv("DB_USER", "postgres")
    db_password = os.getenv("DB_PASSWORD", "postgres")
    db_name = os.getenv("DB_NAME", "budget_tracker")

    return f"postgresql://{db_user}:{db_password}@{db_host}:{db_port}/{db_name}"


# Get database URL from environment
SQLALCHEMY_DATABASE_URL = get_database_url()

# Create engine
engine = create_engine(
    SQLALCHEMY_DATABASE_URL,
    pool_pre_ping=True,  # Verify connections before using them
    pool_size=10,
    max_overflow=20
)

# Create session factory
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

# Base class for models
Base = declarative_base()


def get_db():
    """Dependency for getting database session."""
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()


def init_db():
    """Initialize database tables."""
    Base.metadata.create_all(bind=engine)

